backend/src/chatboxai/requirements.txt
fastapi
uvicorn
requests
python-dotenv
sqlalchemy
pydantic